Renamed package org.inventivetalent.glowapi.
