+++ src/main/java/org/inventivetalent/glowapi/listeners/PlayerJoinListener.java
@@ -1,10 +1,10 @@
-package org.inventivetalent.glow.listeners;
+package org.inventivetalent.glowapi.listeners;
 
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
-import org.inventivetalent.glow.GlowAPI;
+import org.inventivetalent.glowapi.GlowAPI;
 
 public class PlayerJoinListener implements Listener {
