+++ src/main/java/org/inventivetalent/glowapi/packetwrapper/AbstractPacket.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.inventivetalent.glow.packetwrapper;
+package org.inventivetalent.glowapi.packetwrapper;
 
 import java.lang.reflect.InvocationTargetException;
